+++ src/test/particle_filter_tests.cpp
@@ -73,7 +73,7 @@ TEST_F(ParticleFilterTest, DataAssociationTest) {
 TEST_F(ParticleFilterTest, InitTest) {
 
     // Case 1 : Standard deviations set to 0
-    double std_devs[3] = {0.0F, 0.0F, 0.0F};
+    float std_devs[3] = {0.0F, 0.0F, 0.0F};
     mParticleFilter.init(5.0F, 5.0F, 10.0F, std_devs);
     std::vector<Particle> particles = mParticleFilter.particles();
